examples/rag_aiagent_evaluation/requirements.txt
datasets
ragas
llama_index == 0.12.33
langchain-openai
langchain-community
langchain
tabulate
nemoguardrails
trulens== 1.4.9
numpy == 2.2.5
trulens-eval==1.4.9
trulens-apps-llamaindex==1.4.9
trulens-apps-llamaindex==1.4.9
trulens-providers-openai==1.4.9
pandas==2.2.3
duckdb==1.2.2
pydantic==2.11.3
phoenix==0.9.1
arize-phoenix-otel == 0.9.2
openinference-instrumentation-openai == 0.1.27
arize-phoenix-evals == 0.20.6
arize-phoenix == 8.29.0
openinference-instrumentation == 0.1.28